fix: remove more circular deps
