+++ react-components/src/components/RevealCanvas/ViewerContext.tsx
@@ -4,40 +4,19 @@
 import { type DataSourceType, type Cognite3DViewer } from '@cognite/reveal';
 import { createContext, type ReactElement, type ReactNode, useContext } from 'react';
 import { type RevealRenderTarget } from '../../architecture/base/renderTarget/RevealRenderTarget';
-import { type CameraStateParameters, useCameraStateControl } from './hooks/useCameraStateControl';
 
 const ViewerContext = createContext<RevealRenderTarget | null>(null);
 
 export type ViewerContextProviderProps = {
-  cameraState?: CameraStateParameters;
-  setCameraState?: (cameraState?: CameraStateParameters) => void;
   value: RevealRenderTarget | null;
   children: ReactNode;
 };
 
 export const ViewerContextProvider = ({
-  cameraState,
-  setCameraState,
   value,
   children
 }: ViewerContextProviderProps): ReactElement => {
-  return (
-    <ViewerContext.Provider value={value}>
-      <ViewerControls cameraState={cameraState} setCameraState={setCameraState} />
-      {children}
-    </ViewerContext.Provider>
-  );
-};
-
-const ViewerControls = ({
-  cameraState,
-  setCameraState
-}: {
-  cameraState?: CameraStateParameters;
-  setCameraState?: (cameraState?: CameraStateParameters) => void;
-}): ReactNode => {
-  useCameraStateControl(cameraState, setCameraState);
-  return null;
+  return <ViewerContext.Provider value={value}>{children}</ViewerContext.Provider>;
 };
 
 export const useReveal = (): Cognite3DViewer<DataSourceType> => {
